Change the implementation of the `ThreadPool`
`thread::spawn` expects to get some code the thread should run
as soon as the thread is created. However, we want tot create the threads
and have them wait for code that we'll send later.

So `Worker` has defined and the `ThreadPool` has changed to contain workers.
The `Worker` struct is responsible for sending code from the `ThreadPool`
to a `Thread`.